* Preliminary support Multi-faceted Table/Column Definitions.
The idea behind this is that certain definitions have a 'normal' and a
'database' specific setting, and the right one gets automatically used at
runtime.

For example, say we have a binary column which under SQLite and MySQL the
'BLOB' type works fine, but under PostgreSQL we have to use 'BYTEA'.
Our column definition can now be like this:

  [ name => 'bincol', type => 'BLOB', type_pg => 'BYTEA' ]

Internally SQL::DB at connect() time tells the tables what kind of
database we have (based on the dbi: connection string) and the table
objects know which definition to use. 'type_pg' for postgres and 'type' for
everything else.

So far only implemented for things like the default character set and the
mysql ENGINE statement.

SQL::DB
    Next sequence is now seq($name, $count) and the optional $count will
    increment by that amount and return a list of ($count) unique ids.

    Now set table database type at connect() time.

SQL::DB::Row
    Detect and croak() if caller attempt to create a class with duplicate
    column names.

SQL::DB::Schema::Table
    Multifaced definition support for 'type', 'engine', 'default_charset'
    and 'tablespace'.