Make sure that set_*() triggers see all of the incoming values.
When an SQL::DB::Row object is created with the new($hashref) method
the set_col() method is called for each incoming key in the $hashref.
We preset the object with these values in case one of the set_* methods
triggers a subroutine which expects the incoming values to exist in the
object.