Convert SQL::DB::Expr to be ARRAY-based instead of string-based.
This is a rewrite of SQL::DB::Expr to collect string tokens in an
array, rather than concatenating strings together. This is done to
allow quote() and bv() functions to insert quote and bind type objects
into the stream which can be used with DBI->quote or DBI->bind_value
calls just before a query is prepared.